Move mesh to face_3d
As suggested by @nontas
